changed savepoints counter to std::size_t
Changed transaction level (aka savepoints in MySQL DB) from uint
to the std::size_t type.

+++ src/orm/databasemanager.cpp
@@ -177,7 +177,7 @@ bool DatabaseManager::rollbackToSavepoint(const std::size_t id)
     return connection().rollbackToSavepoint(id);
 }
 
-uint DatabaseManager::transactionLevel()
+size_t DatabaseManager::transactionLevel()
 {
     return connection().transactionLevel();
 }
